[MTGOSDK/lib/ScubaDiver] Separate endpoints into partial class
