+++ MTGOSDK/lib/ScubaDiver/src/Endpoints/Client/MakeRegisterClientResponse.cs
@@ -1,0 +1,81 @@
+/** @file
+  Copyright (c) 2021, Xappy.
+  Copyright (c) 2024, Cory Bennett. All rights reserved.
+  SPDX-License-Identifier: Apache-2.0
+**/
+
+using System;
+using System.Collections;
+using System.Collections.Concurrent;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.Linq;
+using System.Net;
+using System.Reflection;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.Diagnostics.Runtime;
+using Newtonsoft.Json;
+
+using MTGOSDK.Core.Compiler.Snapshot;
+using MTGOSDK.Core.Logging;
+using MTGOSDK.Core.Reflection;
+using MTGOSDK.Core.Reflection.Extensions;
+using MTGOSDK.Core.Reflection.Types;
+using MTGOSDK.Core.Remoting.Interop;
+using MTGOSDK.Core.Remoting.Interop.Interactions;
+using MTGOSDK.Core.Remoting.Interop.Interactions.Callbacks;
+using MTGOSDK.Core.Remoting.Interop.Interactions.Client;
+using MTGOSDK.Core.Remoting.Interop.Interactions.Dumps;
+using MTGOSDK.Core.Remoting.Interop.Interactions.Object;
+
+
+namespace ScubaDiver;
+
+public partial class Diver : IDisposable
+{
+  public object _registeredPidsLock = new();
+  public List<int> _registeredPids = new();
+
+  private string MakeRegisterClientResponse(HttpListenerRequest arg)
+  {
+    string pidString = arg.QueryString.Get("process_id");
+    if (pidString == null || !int.TryParse(pidString, out int pid))
+    {
+      return QuickError("Missing parameter 'process_id'");
+    }
+    lock (_registeredPidsLock)
+    {
+      _registeredPids.Add(pid);
+    }
+    Log.Debug("[Diver] New client registered. ID = " + pid);
+    return "{\"status\":\"OK'\"}";
+  }
+
+  private string MakeUnregisterClientResponse(HttpListenerRequest arg)
+  {
+    string pidString = arg.QueryString.Get("process_id");
+    if (pidString == null || !int.TryParse(pidString, out int pid))
+    {
+      return QuickError("Missing parameter 'process_id'");
+    }
+    bool removed;
+    int remaining;
+    lock (_registeredPidsLock)
+    {
+      removed = _registeredPids.Remove(pid);
+      remaining = _registeredPids.Count;
+    }
+    Log.Debug("[Diver] Client unregistered. ID = " + pid);
+
+    UnregisterClientResponse ucResponse = new()
+    {
+      WasRemoved = removed,
+      OtherClientsAmount = remaining
+    };
+
+    return JsonConvert.SerializeObject(ucResponse);
+  }
+}
